line-join-line: fix a bug when joining lines
When lines LINE1 and LINE2 were joined, then cursors from LINE2 were detached
and attached to LINE1 and position was incremented by the LINE1 length (so the
cursor position was the same relative to LINE2 elements). Then contents of LINE1
were concatenated with contents of LINE2.

Unless the cursor position is 0, increasing its position by LINE1 length signals
an error, because LINE1 is not yet concatenated with LINE2, so it has exactly
LENGTH elements (i.e not LENGTH1 + LENGTH2).

This commit addresses this issue by changing line1 contents before cursors are
moved from LINE2 to LINE1.