cursor: attach-cursor when the initarg line is supplied
Previously supplying the argument did associate the line with the cursor but
it did not push the cursor to the line's maintained cursors. That have lead to
an issue where line (when modified) did not update the cursor disregarding of
whether it was left- or right- sticky.

Another issue was that the cursor-position was initialized when the initrarg
:cursor-position was supplied disregarding of whether it was attached to the
line or not, and disregarding of whether it did fit in the line or not -
without signaling appropriate error.

The last issue was that the initarg :cursor-position was not documented.